+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Changelog for NeoFS Node
 ### Fixed
 
 ### Changed
+- Move `fschain_autodeploy` into `fschain.disable_autodeploy` in IR config (#3619)
+- Move `without_mainnet` into `mainnet.enabled` in IR config (#3619)
+- Move `governance.disable` into `mainnet.disable_governance_sync` in IR config (#3619)
+- Move `fee.main_chain` into `mainnet.extra_fee` in IR config (#3619)
+- Move `contracts` into `mainnet.contracts` in IR config (#3619)
 
 ### Removed
 
@@ -18,6 +23,16 @@ Changelog for NeoFS Node
 ### Updating from v0.49.1
 Erasure coding is available in experimental mode. To enable it, set
 `experimental.allow_ec` flag in IR config.
+
+Use IR configuration options:
+- `fschain.disable_autodeploy` instead of deprecated `fschain_autodeploy` with reverted value,
+by default autodeploy is enabled now;
+- `mainnet.enabled` instead of deprecated `without_mainnet` with reverted value,
+by default mainnet is disabled now;
+- `mainnet.disable_governance_sync` instead of deprecated `governance.disable`;
+- `mainnet.extra_fee` instead of deprecated `fee.main_chain`;
+- `mainnet.contracts` instead of deprecated `contracts`;
+Old options are still supported but will be removed in future releases.
 
 ## [0.49.1] - 2025-10-15
 
@@ -2695,8 +2710,7 @@ NeoFS-API v2.0 support and updated brand-new storage node application.
 
 First public review release.
 
-[Unreleased]: https://github.com/nspcc-dev/neofs-node/compare/v0.49.1...master
-[0.49.1]: https://github.com/nspcc-dev/neofs-node/compare/v0.49.0...v0.49.1
+[Unreleased]: https://github.com/nspcc-dev/neofs-node/compare/v0.49.0...master
 [0.49.0]: https://github.com/nspcc-dev/neofs-node/compare/v0.48.3...v0.49.0
 [0.48.3]: https://github.com/nspcc-dev/neofs-node/compare/v0.48.2...v0.48.3
 [0.48.2]: https://github.com/nspcc-dev/neofs-node/compare/v0.48.1...v0.48.2